This is my attempt at a password generator
It uses a for loop to call a function to get input from the user. The
for loop calls a list containing [lower, upper, digits, special] so the
question can be asked 4 times, 1 for each element in the list

The input is added to a dictionary:
key = lower, upper, digits, special
value = number of characters selected by user

For loop cycles through keys, nested for loop generates random value based
on number of characters selected by user. Random characters are added to a
loop. Loops is then randomized and password is displayed to user.